Restricted some actions to only logged in users

+++ components/Like.jsx
@@ -5,6 +5,8 @@ import FavoriteIcon from "@mui/icons-material/Favorite";
 import { useEffect, useState } from "react";
 import { useRouter } from "next/navigation";
 import useSWR, { mutate } from "swr";
+import { useSession } from "next-auth/react";
+import { alertUserToSign } from "@/lib/alertUserToSignIn";
 
 const getLikes = async () => {
   try {
@@ -19,6 +21,7 @@ const getLikes = async () => {
 };
 
 export const Like = ({ item }) => {
+  const { data: session } = useSession();
   const { data, error, isLoading } = useSWR("/api/likes", getLikes);
   let likes = [];
   if (!isLoading && !error) {
@@ -37,29 +40,33 @@ export const Like = ({ item }) => {
 
   //Handle post like to the database.
   const postLike = async (likedPost) => {
-    try {
-      const res = await fetch("/api/likes", {
-        method: "POST",
-        headers: {
-          "Content-type": "application/json",
-        },
-        body: JSON.stringify({
-          likedRecipeId: likedPost._id,
-          likedRecipeUserId: likedPost.userId,
-        }),
-      });
-      const { results } = res.json();
-      console.log(results);
-      if (res.ok) {
-        mutate("/api/likes");
-        setLikedItems(likes);
-        //alertAfterPost();
-        router.push("/recipes");
-      } else {
-        throw new Error("Failed to post a like to this post");
+    if (session) {
+      try {
+        const res = await fetch("/api/likes", {
+          method: "POST",
+          headers: {
+            "Content-type": "application/json",
+          },
+          body: JSON.stringify({
+            likedRecipeId: likedPost._id,
+            likedRecipeUserId: likedPost.userId,
+          }),
+        });
+        const { results } = res.json();
+        console.log(results);
+        if (res.ok) {
+          mutate("/api/likes");
+          setLikedItems(likes);
+          //alertAfterPost();
+          router.push("/recipes");
+        } else {
+          throw new Error("Failed to post a like to this post");
+        }
+      } catch (error) {
+        console.log(error);
       }
-    } catch (error) {
-      console.log(error);
+    } else {
+      alertUserToSign();
     }
   };
   //Handle deleting like from the database
@@ -71,14 +78,18 @@ export const Like = ({ item }) => {
         reci.likedRecipeUserId === likedPost.userId
     );
 
-    if (confirmed) {
-      const res = await fetch(`/api/likes?id=${recipeToDelete._id}`, {
-        method: "DELETE",
-      });
+    if (session) {
+      if (confirmed) {
+        const res = await fetch(`/api/likes?id=${recipeToDelete._id}`, {
+          method: "DELETE",
+        });
 
-      if (res.ok) {
-        alert("You just disliked this post");
+        if (res.ok) {
+          alert("You just disliked this post");
+        }
       }
+    } else {
+      alertUserToSign();
     }
   };
